requirements.txt
apache-airflow==1.10.11
beautifulsoup4==4.9.1
folium==0.10
matplotlib==3.2.2
numpy==1.19.0
pandas==1.0.5
requests==2.22.0
requests-unixsocket==0.2.0
seaborn==0.10.1
